Autentikointi poistolle

+++ controllers/blogs.js
@@ -33,6 +33,12 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
     const id = request.params.id
 
     await Blog.findByIdAndRemove(id)
